splitting code into components
