+++ src/App.jsx
@@ -1,75 +1,14 @@
-import React, { useState } from "react";
-
 import Header from "./components/Header/Header";
-import CoreConcepts from "./components/CoreConcepts";
-import TabButton from "./components/TabButton";
-
-import { CORE_CONCEPTS } from "./data";
-import { EXAMPLES } from "./data";
+import CoreConcept from "./components/CoreConcept";
+import Examples from "./components/Examples";
 
 function App() {
-  const [selectedTopic, setSelectedTopic] = useState();
-  //rendering content conditionally
-  let tabContent = <p>Please Select a Topic</p>;
-  if (selectedTopic) {
-    tabContent = (
-      <div id="tab-content">
-        <h3>{EXAMPLES[selectedTopic].title}</h3>
-        <p>{EXAMPLES[selectedTopic].description}</p>
-        <pre>
-          <code>{EXAMPLES[selectedTopic].code}</code>
-        </pre>
-      </div>
-    );
-  }
-
-  const handleSelect = (selectedButton) => {
-    setSelectedTopic(selectedButton);
-    console.log(selectedTopic);
-  };
-
   return (
     <>
       <Header />
       <main>
-        <section id="core-concepts">
-          <h2>Core Concepts</h2>
-          <ul>
-            {CORE_CONCEPTS.map((concept) => (
-              <CoreConcepts key={concept.title} {...concept} />
-            ))}
-          </ul>
-        </section>
-        <section id="examples">
-          <h2>Examples</h2>
-          <menu>
-            <TabButton
-              isSelected={selectedTopic === "components"}
-              onSelect={() => handleSelect("components")}
-            >
-              Components
-            </TabButton>
-            <TabButton
-              isSelected={selectedTopic === "jsx"}
-              onSelect={() => handleSelect("jsx")}
-            >
-              JSX
-            </TabButton>
-            <TabButton
-              isSelected={selectedTopic === "props"}
-              onSelect={() => handleSelect("props")}
-            >
-              Props
-            </TabButton>
-            <TabButton
-              isSelected={selectedTopic === "state"}
-              onSelect={() => handleSelect("state")}
-            >
-              State
-            </TabButton>
-          </menu>
-          {tabContent}
-        </section>
+        <CoreConcept />
+        <Examples />
       </main>
     </>
   );
